backend/pyproject.toml
[project]
name = "gemini-unified-gateway"
version = "0.1.0"
description = "Unified Gemini web gateway with OpenAI-compatible API"
requires-python = ">=3.11"
dependencies = [
    # --- Upstream adapters ---
    "gemini-webapi[browser]>=1.15.0",
    "gemini-web-mcp-cli>=0.1.0",
    # --- Web framework ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "gunicorn>=22.0.0",
    "sse-starlette>=2.1.0",
    "websockets>=13.0",
    "python-multipart>=0.0.12",
    # --- Database ---
    "sqlalchemy[asyncio]>=2.0.35",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # --- Cache / Queue ---
    "redis[hiredis]>=5.2.0",
    "celery[redis]>=5.4.0",
    "flower>=2.0.0",
    "celery-redbeat>=2.2.0",
    "celery-singleton>=0.3.1",
    # --- Auth / Security ---
    "pyjwt[crypto]>=2.9.0",
    "bcrypt>=4.0.1",
    "cryptography>=43.0.0",
    "passlib[bcrypt]>=1.7.4",
    # --- Schemas ---
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.2.0",
    # --- HTTP ---
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "aiohttp>=3.10.0",
    # --- Logging / Metrics ---
    "structlog>=24.4.0",
    "loguru>=0.7.2",
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    # --- CLI ---
    "typer[all]>=0.12.0",
    "rich>=13.9.0",
    # --- MCP ---
    "fastmcp>=2.12.0",
    "mcp>=1.0.0",
    # --- Rate limiting ---
    "slowapi>=0.1.9",
    # --- Utilities ---
    "python-dotenv>=1.0.1",
    "orjson>=3.10.0",
    "tenacity>=9.0.0",
    "cachetools>=5.5.0",
    "croniter>=3.0.0",
    "python-dateutil>=2.9.0",
    "pillow>=10.4.0",
    "browser-cookie3>=0.19.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-xdist>=3.5.0",
    "factory-boy>=3.3.0",
    "faker>=30.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "sqlalchemy[mypy]>=2.0.35",
    "pre-commit>=4.0.0",
    "bandit>=1.7.10",
    "locust>=2.31.0",
    "watchfiles>=0.24.0",
]

[project.scripts]
app = "cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "cli*"]
exclude = ["alembic*", "tests*"]

[tool.ruff]
target-version = "py311"
line-length = 120
select = ["E","W","F","I","N","UP","S","B","A","C4","PT","RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short -q"
